moved help_action into parameters.options since a module dedicated specifically to it was not necessary
