+++ lib/cli_command_parser/parameters/__init__.py
@@ -1,6 +1,6 @@
 from .base import Parameter, BasePositional, BaseOption
 from .choice_map import SubCommand, Action
 from .groups import ParamGroup
-from .options import Option, Flag, Counter, ActionFlag, action_flag, before_main, after_main, TriFlag
+from .options import Option, Flag, Counter, ActionFlag, action_flag, before_main, after_main, TriFlag, help_action
 from .pass_thru import PassThru
 from .positionals import Positional
