+++ tests/test_parameters/test_action_flags.py
@@ -5,9 +5,8 @@ from unittest import main
 from unittest.mock import Mock, seal
 
 from cli_command_parser import Command, Action, no_exit_handler, ActionFlag, ParamGroup
-from cli_command_parser.actions import help_action
 from cli_command_parser.context import Context
-from cli_command_parser.parameters import before_main, after_main, action_flag
+from cli_command_parser.parameters import before_main, after_main, action_flag, help_action
 from cli_command_parser.exceptions import CommandDefinitionError, ParameterDefinitionError, ParamConflict
 from cli_command_parser.testing import ParserTest, RedirectStreams
